// javascript/src/openapi/models/CursorPageInvoice.ts
/**
 * Pay OpenAPI Documentations
 * This is a pay service api server.
 *
 * OpenAPI spec version: 1.0
 * Contact: support@openweb3.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Invoice } from "./Invoice";

export class CursorPageInvoice {
  static readonly discriminator: string | undefined = undefined;
  static readonly attributeTypeMap: Array<{ name: string; baseName: string; type: string; format: string }> = [
    {
      "name": "items",
      "baseName": "items",
      "type": "Array<Invoice>",
      "format": "",
    },
    {
      "name": "nextCursor",
      "baseName": "next_cursor",
      "type": "string",
      "format": "",
    },
    {
      "name": "prevCursor",
      "baseName": "prev_cursor",
      "type": "string",
      "format": "",
    }];
  "items"?: Array<Invoice>;
  "nextCursor"?: string;
  "prevCursor"?: string;

  static getAttributeTypeMap() {
    return CursorPageInvoice.attributeTypeMap;
  }
}